+++ services/kuber/server/domain/usecases/patch_nodes.go
@@ -1,0 +1,23 @@
+package usecases
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/berops/claudie/internal/utils"
+	"github.com/berops/claudie/proto/pb"
+	"github.com/berops/claudie/services/kuber/server/domain/utils/nodes"
+)
+
+func (u *Usecases) PatchNodes(ctx context.Context, request *pb.PatchNodeTemplateRequest) (*pb.PatchNodeTemplateResponse, error) {
+	logger := utils.CreateLoggerWithClusterName(utils.GetClusterID(request.Cluster.ClusterInfo))
+
+	patcher := nodes.NewPatcher(request.Cluster)
+	if err := patcher.PatchProviderID(logger); err != nil {
+		logger.Err(err).Msgf("Error while patching nodes")
+		return nil, fmt.Errorf("error while patching nodes for %s : %w", request.Cluster.ClusterInfo.Name, err)
+	}
+
+	logger.Info().Msgf("Nodes were successfully patched")
+	return &pb.PatchNodeTemplateResponse{}, nil
+}
