+++ services/kuber/server/domain/utils/nodes/delete.go
