+++ services/kuber/server/domain/utils/utils.go
@@ -1,22 +1,22 @@
-package main
+package utils
 
 import (
 	"fmt"
 
 	"github.com/berops/claudie/internal/utils"
 	"github.com/berops/claudie/proto/pb"
-	"github.com/berops/claudie/services/kuber/server/secret"
+	"github.com/berops/claudie/services/kuber/server/domain/utils/secret"
 )
 
-type outputType string
+type OutputType string
 
 const (
-	kubeconfigSecret outputType = "kubeconfig"
-	metadataSecret   outputType = "metadata"
+	KubeconfigSecret OutputType = "kubeconfig"
+	MetadataSecret   OutputType = "metadata"
 )
 
 // getSecretMetadata returns metadata for secrets created in the management cluster as a Claudie output.
-func getSecretMetadata(ci *pb.ClusterInfo, projectName string, outputType outputType) secret.Metadata {
+func GetSecretMetadata(ci *pb.ClusterInfo, projectName string, outputType OutputType) secret.Metadata {
 	cid := utils.GetClusterID(ci)
 	return secret.Metadata{
 		Name: fmt.Sprintf("%s-%s", cid, outputType),
